JarvisAssistant/requirements.txt
google-generativeai
mss
pyttsx3
SpeechRecognition
pyaudio
pynput
pillow
python-dotenv
